# Replace AsByteSliceMut with Fill