**Breaking:** this would be the first step towards 0.8.

Note that once `specialization` is available, `Fill` becomes redundant. Until then, we need it (in some form).

Benchmark results:
```
test gen_1kb_u16_fill        ... bench:         325 ns/iter (+/- 20) = 3150 MB/s
test gen_1kb_u16_gen_array   ... bench:         876 ns/iter (+/- 80) = 1168 MB/s
test gen_1kb_u16_iter_repeat ... bench:       1,302 ns/iter (+/- 112) = 786 MB/s
test gen_1kb_u16_sample_iter ... bench:       1,251 ns/iter (+/- 63) = 818 MB/s
test gen_1kb_u64_fill        ... bench:         331 ns/iter (+/- 12) = 3093 MB/s
test gen_1kb_u64_gen_array   ... bench:         236 ns/iter (+/- 11) = 4338 MB/s
test gen_1kb_u64_iter_repeat ... bench:         335 ns/iter (+/- 11) = 3056 MB/s
test gen_1kb_u64_sample_iter ... bench:         330 ns/iter (+/- 8) = 3103 MB/s
```
Without opt:
```
test gen_1kb_u16_fill        ... bench:         484 ns/iter (+/- 99) = 2115 MB/s
test gen_1kb_u64_fill        ... bench:         460 ns/iter (+/- 19) = 2226 MB/s
```
With `AsByteSliceMut`:
```
test gen_1kb_u16_fill        ... bench:         337 ns/iter (+/- 30) = 3038 MB/s
test gen_1kb_u64_fill        ... bench:         327 ns/iter (+/- 13) = 3131 MB/s
```

Also, add `const_generics` feature flag. See also #939.