@dhardy 2020-02-29
Looks like we'll have to remove even feature-gated `const_generics` until the MSRV is increased.

Other than that, review @vks?