@dhardy 2020-03-06
I removed the const-generics impl since even behind a feature flag it was erroring out the 1.32 build. The code still lives [here](https://github.com/rust-random/rand/tree/nightly-const-generics).

Still waiting on a review, but maybe it's time to give up on that.